Updated to conform closer to proftpd standards and correct an issue navigating directories
Changes for #1
Now is closer to proftpd standards.
There was a problem I found when performing additional testing for #1 that this corrects.
Specifically when the folder in the middle of the path is a different case compared to on-disk, it was not handling it correctly.
e.g. given the valid path `/V2.4/mod_case/code/READ_ME.txt`  if the command is issued with `/V2.4/Mod_case/code/READ_ME.txt`, after it drilled up to find the correct casing of `mod_case`, it would go back down to to find the correct case of `code` and discover that it is now an exact match, then kick out, and then not discover the fully corrected path.